requirements.txt
flask==2.3.2
python-docx==0.8.11
python-dotenv==1.0.0
pywin32==306; sys_platform == 'win32'
pytest==7.4.0
pytest-cov==4.1.0
python-pptx==0.6.21
email-validator==2.0.0.post2
flask-wtf==1.1.1
libreoffice-headless==7.5.0; sys_platform == 'linux'